refactor: 게임 관련 컴포넌트 games 폴더로 이동
