+++ src/pages/games/[id]/GamePage.js
@@ -1,9 +1,9 @@
 import Component from '../../../core/Component';
 import styles from './GamePage.module.css';
-import GameHeader from '../../../components/game-header/GameHeader';
-import GameEditor from '../../../components/game-editor/GameEditor';
-import GamePreview from '../../../components/game-preview/GamePreview';
-import GameOtherCode from '../../../components/game-other-code/GameOtherCode';
+import GameHeader from '../../../components/games/game-header/GameHeader';
+import GameEditor from '../../../components/games/game-editor/GameEditor';
+import GamePreview from '../../../components/games/game-preview/GamePreview';
+import GameOtherCode from '../../../components/games/game-other-code/GameOtherCode';
 import Store from '../../../store/Store';
 import escapeHtml from '../../../utils/escapeHtml';
 import htmlGameData from '../../../data/games/htmlGames.json';
